automate_bindings: basic support of overloading
Overloaded methods are distinguished by the number of arguments only.

Function getConstructor was replaced with getConstructors.

A group of functios (storeArgc, ifArgc, elseArgc, elseFail) was added to
implement dispatching by number of arguments. If overloaded functions have
the same number of arguments only one of them (arbitrary one) is used.
Ideally we should also take types of arguments into account to distinguish
overloaded functions with the same number of arguments. But current approach
should work for the most of cases and it should produce compilable code.
Previous version produced broken code because of multiple Lua C functions
with the same name.