Add WTestEnvironment to rockspec, globals and init
We have to comment it in rockspec and init.cpp because it depends on
wttest, which is incompatible with wthttp (which is needed for WServer).
Anyway, the main usage of this module is testing so usually we dont need
both WServer and WTestEnvironment.
To activate WTestEnvironment just uncomment it in .rockspec (do not
forget to uncomment wttest in the list of libraries as well) and init.cpp.
Then comment WServer module and erase/comment wthhtp in the list of
libraries.