automate_bindings.py: fix bug in findCorrespondingKeyInDict()
In case of specific types (e.g. enums) the full key may contain
index key by accident (e.g. WPaintedDevice <--> int).
Anyway, we're able to prevent it, because if key's specific, it's
stored as exactly the same in dict, hence we can just check
if this key exists before searching for partial matching.